+++ src/SettingsList.h
@@ -402,40 +402,41 @@ inline std::vector<SettingInfo> getSettingsList(const SdCardFontRegistry* regist
                           {StrId::STR_LONG_PRESS_BEHAVIOR_OFF, StrId::STR_LONG_PRESS_BEHAVIOR_SKIP,
                            StrId::STR_LONG_PRESS_BEHAVIOR_ORIENTATION},
                           "longPressButtonBehavior", StrId::STR_CAT_CONTROLS));
-    add(SettingInfo::Enum(
-            StrId::STR_SHORT_PWR_BTN, &CrossPointSettings::shortPwrBtn,
-            {StrId::STR_IGNORE, StrId::STR_SLEEP, StrId::STR_PAGE_TURN, StrId::STR_TOGGLE_BOOKMARK,
-             StrId::STR_READING_STATS, StrId::STR_MARK_FINISHED, StrId::STR_FORCE_REFRESH, StrId::STR_CHANGE_FONT,
-             StrId::STR_TOGGLE_GUIDE_DOTS, StrId::STR_TOGGLE_BIONIC_READING, StrId::STR_CYCLE_PAGE_TURN,
-             StrId::STR_SYNC_PROGRESS, StrId::STR_FILE_TRANSFER, StrId::STR_SCREENSHOT_BUTTON},
-            "shortPwrBtn", StrId::STR_CAT_CONTROLS)
+    add(SettingInfo::Enum(StrId::STR_SHORT_PWR_BTN, &CrossPointSettings::shortPwrBtn,
+                          {StrId::STR_IGNORE, StrId::STR_SLEEP, StrId::STR_PAGE_TURN, StrId::STR_TOGGLE_BOOKMARK,
+                           StrId::STR_READING_STATS, StrId::STR_MARK_FINISHED, StrId::STR_FORCE_REFRESH,
+                           StrId::STR_CHANGE_FONT, StrId::STR_TOGGLE_GUIDE_DOTS, StrId::STR_TOGGLE_BIONIC_READING,
+                           StrId::STR_CYCLE_PAGE_TURN, StrId::STR_SYNC_PROGRESS, StrId::STR_FILE_TRANSFER,
+                           StrId::STR_SCREENSHOT_BUTTON, StrId::STR_READER_DARK_MODE},
+                          "shortPwrBtn", StrId::STR_CAT_CONTROLS)
             .withEnumRawValues({CrossPointSettings::IGNORE, CrossPointSettings::SLEEP, CrossPointSettings::PAGE_TURN,
                                 CrossPointSettings::TOGGLE_BOOKMARK, CrossPointSettings::READING_STATS,
                                 CrossPointSettings::MARK_FINISHED, CrossPointSettings::FORCE_REFRESH,
                                 CrossPointSettings::TOGGLE_FONT, CrossPointSettings::TOGGLE_GUIDE_DOTS,
                                 CrossPointSettings::TOGGLE_BIONIC_READING, CrossPointSettings::CYCLE_PAGE_TURN,
                                 CrossPointSettings::SYNC_PROGRESS, CrossPointSettings::FILE_TRANSFER,
-                                CrossPointSettings::SCREENSHOT}));
-    add(SettingInfo::Enum(
-            StrId::STR_LONG_PRESS_ACTION, &CrossPointSettings::longPwrBtn,
-            {StrId::STR_IGNORE, StrId::STR_SLEEP, StrId::STR_PAGE_TURN, StrId::STR_TOGGLE_BOOKMARK,
-             StrId::STR_READING_STATS, StrId::STR_MARK_FINISHED, StrId::STR_FORCE_REFRESH, StrId::STR_CHANGE_FONT,
-             StrId::STR_TOGGLE_GUIDE_DOTS, StrId::STR_TOGGLE_BIONIC_READING, StrId::STR_CYCLE_PAGE_TURN,
-             StrId::STR_SYNC_PROGRESS, StrId::STR_FILE_TRANSFER, StrId::STR_SCREENSHOT_BUTTON},
-            "longPwrBtn", StrId::STR_CAT_CONTROLS)
+                                CrossPointSettings::SCREENSHOT, CrossPointSettings::TOGGLE_DARK_MODE}));
+    add(SettingInfo::Enum(StrId::STR_LONG_PRESS_ACTION, &CrossPointSettings::longPwrBtn,
+                          {StrId::STR_IGNORE, StrId::STR_SLEEP, StrId::STR_PAGE_TURN, StrId::STR_TOGGLE_BOOKMARK,
+                           StrId::STR_READING_STATS, StrId::STR_MARK_FINISHED, StrId::STR_FORCE_REFRESH,
+                           StrId::STR_CHANGE_FONT, StrId::STR_TOGGLE_GUIDE_DOTS, StrId::STR_TOGGLE_BIONIC_READING,
+                           StrId::STR_CYCLE_PAGE_TURN, StrId::STR_SYNC_PROGRESS, StrId::STR_FILE_TRANSFER,
+                           StrId::STR_SCREENSHOT_BUTTON, StrId::STR_READER_DARK_MODE},
+                          "longPwrBtn", StrId::STR_CAT_CONTROLS)
             .withEnumRawValues({CrossPointSettings::IGNORE, CrossPointSettings::SLEEP, CrossPointSettings::PAGE_TURN,
                                 CrossPointSettings::TOGGLE_BOOKMARK, CrossPointSettings::READING_STATS,
                                 CrossPointSettings::MARK_FINISHED, CrossPointSettings::FORCE_REFRESH,
                                 CrossPointSettings::TOGGLE_FONT, CrossPointSettings::TOGGLE_GUIDE_DOTS,
                                 CrossPointSettings::TOGGLE_BIONIC_READING, CrossPointSettings::CYCLE_PAGE_TURN,
                                 CrossPointSettings::SYNC_PROGRESS, CrossPointSettings::FILE_TRANSFER,
-                                CrossPointSettings::SCREENSHOT}));
-    add(SettingInfo::Enum(StrId::STR_LONG_PRESS_MENU_ACTION, &CrossPointSettings::longPressMenuAction,
-                          {StrId::STR_IGNORE, StrId::STR_SLEEP, StrId::STR_TOGGLE_BOOKMARK, StrId::STR_READING_STATS,
-                           StrId::STR_MARK_FINISHED, StrId::STR_FORCE_REFRESH, StrId::STR_CHANGE_FONT,
-                           StrId::STR_TOGGLE_GUIDE_DOTS, StrId::STR_TOGGLE_BIONIC_READING, StrId::STR_CYCLE_PAGE_TURN,
-                           StrId::STR_SYNC_PROGRESS, StrId::STR_FILE_TRANSFER, StrId::STR_SCREENSHOT_BUTTON},
-                          "longPressMenuAction", StrId::STR_CAT_CONTROLS)
+                                CrossPointSettings::SCREENSHOT, CrossPointSettings::TOGGLE_DARK_MODE}));
+    add(SettingInfo::Enum(
+            StrId::STR_LONG_PRESS_MENU_ACTION, &CrossPointSettings::longPressMenuAction,
+            {StrId::STR_IGNORE, StrId::STR_SLEEP, StrId::STR_TOGGLE_BOOKMARK, StrId::STR_READING_STATS,
+             StrId::STR_MARK_FINISHED, StrId::STR_FORCE_REFRESH, StrId::STR_CHANGE_FONT, StrId::STR_TOGGLE_GUIDE_DOTS,
+             StrId::STR_TOGGLE_BIONIC_READING, StrId::STR_CYCLE_PAGE_TURN, StrId::STR_SYNC_PROGRESS,
+             StrId::STR_FILE_TRANSFER, StrId::STR_SCREENSHOT_BUTTON, StrId::STR_READER_DARK_MODE},
+            "longPressMenuAction", StrId::STR_CAT_CONTROLS)
             .withEnumRawValues(
                 {CrossPointSettings::LONG_MENU_OFF, CrossPointSettings::LONG_MENU_SLEEP,
                  CrossPointSettings::LONG_MENU_TOGGLE_BOOKMARK, CrossPointSettings::LONG_MENU_READING_STATS,
@@ -443,7 +444,7 @@ inline std::vector<SettingInfo> getSettingsList(const SdCardFontRegistry* regist
                  CrossPointSettings::LONG_MENU_CHANGE_FONT, CrossPointSettings::LONG_MENU_TOGGLE_GUIDE_DOTS,
                  CrossPointSettings::LONG_MENU_TOGGLE_BIONIC, CrossPointSettings::LONG_MENU_CYCLE_PAGE_TURN,
                  CrossPointSettings::LONG_MENU_SYNC_PROGRESS, CrossPointSettings::LONG_MENU_FILE_TRANSFER,
-                 CrossPointSettings::LONG_MENU_SCREENSHOT}));
+                 CrossPointSettings::LONG_MENU_SCREENSHOT, CrossPointSettings::LONG_MENU_TOGGLE_DARK_MODE}));
 
     // --- System ---
     add(SettingInfo::Value(